pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wechat-stream-ocr"
version = "0.1.0"
description = "WebSocket screenshot diffing and OCR pipeline for WeChat chat windows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "Pillow",
  "pyzmq",
  "websockets",
]

[project.optional-dependencies]
ocr = [
  "paddlepaddle",
  "paddleocr",
  "paddlex[ocr]",
]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
wechat-stream-ocr = "wechat_stream_ocr.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
